+++ tools/ttp/sizedir.do.pl
@@ -39,6 +39,7 @@ use File::Path qw( remove_tree );
 use File::Find;
 
 use TTP::Metric;
+use TTP::Path;
 
 my $defaults = {
 	help => 'no',
@@ -193,7 +194,7 @@ $count += 1 if $opt_dircmd;
 msgErr( "one of '--dirpath' and '--dircmd' options must be specified" ) if $count != 1;
 
 # if we have a source cmd, get the path and make it exist to be sure to have something to publish
-$opt_dirpath = TTP::fromCommand( $opt_dircmd, { makeExist => true }) if $opt_dircmd;
+$opt_dirpath = TTP::Path::( $opt_dircmd, { makeExist => true }) if $opt_dircmd;
 
 if( !TTP::errs()){
 	doComputeSize();